move the Buffer class to a submodule called tip
